Additional test coverage and minor refactor
