+++ util/spec_helper.rb
@@ -68,6 +68,7 @@ require_relative '../pages/page'
 require_relative '../pages/page'
 require_relative '../pages/oec/blue_page'
 require_relative '../pages/cal_net_page'
+require_relative '../pages/canvas/canvas_people_page'
 require_relative '../pages/canvas/canvas_page'
 require_relative '../pages/canvas/canvas_assignments_page'
 require_relative '../pages/canvas/canvas_announce_discuss_page'
